Do not delete python shared libraries outside __pyccel__ or __epyccel__ folders by default

+++ pyccel/commands/pyccel_clean.py
@@ -11,7 +11,7 @@ import sysconfig
 
 ext_suffix = sysconfig.get_config_var('EXT_SUFFIX')
 
-def pyccel_clean(path_dir = None, recursive = True):
+def pyccel_clean(path_dir = None, recursive = True, remove_shared_libs = False):
     """ Remove __pyccel__ and __epyccel__ folders as well
     as any python shared libraries from the directory path_dir
 
@@ -35,5 +35,5 @@ def pyccel_clean(path_dir = None, recursive = True):
             shutil.rmtree( file_name, ignore_errors=True)
         elif not os.path.isfile(file_name) and recursive:
             pyccel_clean(file_name)
-        elif f.endswith(ext_suffix):
+        elif remove_shared_libs and f.endswith(ext_suffix):
             os.remove(file_name)
